+++ demosplan/DemosPlanCoreBundle/Controller/Survey/SurveyController.php
@@ -19,8 +19,8 @@ use demosplan\DemosPlanCoreBundle\Logic\Survey\SurveyCreateHandler;
 use demosplan\DemosPlanCoreBundle\Logic\Survey\SurveyHandler;
 use demosplan\DemosPlanCoreBundle\Logic\Survey\SurveyNewHandler;
 use demosplan\DemosPlanCoreBundle\Logic\Survey\SurveyUpdateHandler;
+use demosplan\DemosPlanCoreBundle\Validator\SurveyValidator;
 use demosplan\DemosPlanProcedureBundle\Logic\ProcedureHandler;
-use demosplan\DemosPlanSurveyBundle\Validator\SurveyValidator;
 use Exception;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
