+++ demosplan/DemosPlanCoreBundle/Controller/Survey/SurveyVoteAPIController.php
@@ -26,8 +26,8 @@ use demosplan\DemosPlanCoreBundle\Logic\Survey\SurveyVoteCreateHandler;
 use demosplan\DemosPlanCoreBundle\Logic\Survey\SurveyVoteHandler;
 use demosplan\DemosPlanCoreBundle\Logic\Survey\SurveyVoteService;
 use demosplan\DemosPlanCoreBundle\ResourceTypes\SurveyVoteResourceType;
+use demosplan\DemosPlanCoreBundle\Validator\SurveyVoteValidator;
 use demosplan\DemosPlanProcedureBundle\Logic\CurrentProcedureService;
-use demosplan\DemosPlanSurveyBundle\Validator\SurveyVoteValidator;
 use Exception;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
