+++ demosplan/DemosPlanCoreBundle/Validator/SurveyVoteValidator.php
@@ -8,7 +8,7 @@
  * All rights reserved
  */
 
-namespace demosplan\DemosPlanSurveyBundle\Validator;
+namespace demosplan\DemosPlanCoreBundle\Validator;
 
 use DateTime;
 use DemosEurope\DemosplanAddon\Validator\JsonSchemaValidator;
