+++ tests/backend/core/Survey/Unit/ProcedureMatchSurveyTest.php
@@ -13,7 +13,7 @@ namespace Tests\Core\Survey\Unit;
 use DateTime;
 use demosplan\DemosPlanCoreBundle\DataFixtures\ORM\TestData\LoadSurveyData;
 use demosplan\DemosPlanCoreBundle\Exception\SurveyInputDataException;
-use demosplan\DemosPlanSurveyBundle\Validator\SurveyValidator;
+use demosplan\DemosPlanCoreBundle\Validator\SurveyValidator;
 use Exception;
 use InvalidArgumentException;
