+++ tests/backend/core/Survey/Unit/SurveyValidJsonTest.php
@@ -12,7 +12,7 @@ namespace Tests\Core\Survey\Unit;
 
 use DemosEurope\DemosplanAddon\Utilities\Json;
 use demosplan\DemosPlanCoreBundle\Exception\SurveyInputDataException;
-use demosplan\DemosPlanSurveyBundle\Validator\SurveyValidator;
+use demosplan\DemosPlanCoreBundle\Validator\SurveyValidator;
 use Exception;
 use JsonSchema\Exception\InvalidSchemaException;
